bilet

+++ bilet.cpp
@@ -3,9 +3,9 @@
 
 using namespace std;
 
-long int find_max_subarray(int A[], int n) {
-    long int current_sum = 0;
-    long int max = 0;
+long long int find_max_subarray(int A[], int n) {
+    long long int current_sum = 0;
+    long long int max = 0;
     for (int i = 0; i < n; i++) {
         current_sum += A[i];
         if (current_sum < 0)
@@ -18,20 +18,20 @@ long int find_max_subarray(int A[], int n) {
     return max;
 }
 
-long int find_min_subarray(int A[], int n) {
+long long int find_min_subarray(int A[], int n) {
     for (int i = 0; i < n; i++)
         A[i] = -A[i];
 
     return -find_max_subarray(A, n);
 }
 
-long int maximize_profit(int A[], int n) {
-    long int sum = 0;
+long long int maximize_profit(int A[], int n) {
+    long long int sum = 0;
     for (int i = 0; i < n; i++)
         sum += A[i];
 
-    long int max_subarray = find_max_subarray(A, n);
-    long int min_subarray = find_min_subarray(A, n);
+    long long int max_subarray = find_max_subarray(A, n);
+    long long int min_subarray = find_min_subarray(A, n);
 
     if (sum + max_subarray > sum - min_subarray)
         return sum + max_subarray;
@@ -40,7 +40,7 @@ long int maximize_profit(int A[], int n) {
 
 }
 
-long int solution() {
+long long int solution() {
     int n;
     cin >> n;
     n--;
